mysite/requirements.txt
# Core Django and related packages
Django==4.1.12
djangorestframework==3.14.0
django-environ==0.10.0
asgiref==3.6.0
sqlparse==0.4.3
pytz==2023.3

# Database drivers
pymysql==1.0.3
mysqlclient==2.1.1

# HTTP and networking
requests==2.31.0
urllib3==1.26.18
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4

# Data processing (compatible with Django 4.1)
pandas==2.0.3
numpy==1.24.3
python-dateutil==2.8.2

# Testing
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
coverage==7.2.7

# Development tools
click==8.1.7
rich==13.4.2
typer==0.9.0

# Security
cryptography==41.0.4

# Utilities
six==1.16.0
packaging==23.1
typing_extensions==4.7.1
